requirements.txt
transformers>=4.44.0
accelerate>=0.30.0
trl>=0.8.6
peft>=0.11.1
datasets>=2.19.0
evaluate>=0.4.2
scikit-learn>=1.4.0
numpy>=1.26.0
pyyaml>=6.0.1
